Moved props and created TitleHero
